refactor: Combine testCases and test code
